Refactor: navigate to folder page if token exists
