+++ src/components/SignForm.module.css
@@ -36,13 +36,41 @@
 .form {
   display: flex;
   flex-direction: column;
-  gap: 12px;
+  gap: 32px;
   position: relative;
 }
 
 .eye {
   position: absolute;
   right: 15px;
-  top: 58%;
+  top: 50px;
   cursor: pointer;
+}
+
+.inputBox {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  gap: 12px;
+}
+
+.button {
+  display: flex;
+  width: 100%;
+  padding: 16px 20px;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  border: none;
+  border-radius: 8px;
+  background: linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+  color: #f5f5f5;
+  font-size: 18px;
+  font-weight: 600;
+  cursor: pointer;
+}
+
+.button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed;
 }
